+++ packages/sku/src/openBrowser.ts
@@ -3,16 +3,16 @@
 
 import chalk from 'chalk';
 import open from 'open';
-import isCI from '../utils/isCI.js';
+import isCI from './utils/isCI.js';
 import getDefaultBrowser from 'default-browser';
 import { fileURLToPath } from 'node:url';
 import { resolve } from 'node:path';
 import debug from 'debug';
-import { execAsync } from '../utils/execAsync.js';
+import { execAsync } from './utils/execAsync.js';
 
 const log = debug('sku:openBrowser');
 
-export const BIN_DIR = resolve(fileURLToPath(import.meta.url), '../../../bin');
+const BIN_DIR = resolve(fileURLToPath(import.meta.url), '../../bin');
 
 const OSX_CHROME = 'google chrome';