+++ packages/sku/src/program/commands/init/init.action.ts
@@ -23,7 +23,7 @@ import {
   isEmptyDir,
   banner,
 } from '@sku-lib/utils';
-import { write as prettierWrite } from '../../../services/prettier/runPrettier.js';
+import { write as prettierWrite } from '../../../services/prettier.js';
 import { fix as esLintFix } from '../../../services/eslint/runESLint.js';
 import install from '../../../services/packageManager/install.js';