+++ packages/sku/src/services/prettier.ts
@@ -1,16 +1,16 @@
-import exists from '../../utils/exists.js';
+import exists from '../utils/exists.js';
 import path, { dirname } from 'node:path';
 import chalk from 'chalk';
-import { runBin } from '../../utils/runBin.js';
+import { runBin } from '../utils/runBin.js';
 import { getPathFromCwd } from '@sku-lib/utils';
-import { suggestScript } from '../../utils/suggestScript.js';
+import { suggestScript } from '../utils/suggestScript.js';
 
 import { fileURLToPath } from 'node:url';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
 const prettierIgnorePath = getPathFromCwd('.prettierignore');
-const prettierConfigPath = path.join(__dirname, '../../config/prettier.js');
+const prettierConfigPath = path.join(__dirname, '../config/prettier.js');
 
 const runPrettier = async ({
   write,